Added edit clubs page
