+++ client/src/components/clubs/club-popup.scss
@@ -124,148 +124,8 @@
     margin-bottom: 1rem;
 }
 
-/*
- * ##########################
- * ##### EDIT CLUB INFO #####
- * ##########################
- */
-
 .club-popup-edit-button {
     margin-bottom: 0.5rem;
-}
-
-.club-popup-edit {
-    text-align: left;
-}
-
-// Container is relative because absolute needs a non-static positioned ancestor
-.club-popup-edit-image-container {
-    position: relative;
-    width: 100%;
-    height: 10.5rem;
-}
-
-// Input (button) of type 'file' to upload images
-.club-popup-cover-photo-upload {
-    position: absolute;
-    bottom: 0.5rem;
-    left: 0.5rem;
-}
-
-// Everything underneath the cover photo upload
-// This container has a padding that applies to all the 'edit' elements
-.club-popup-edit-bottom {
-    padding: 0.75rem;
-}
-
-.club-popup-name-advised-div {
-    display: flex;
-    flex-direction: row;
-    justify-content: left;
-}
-
-.club-popup-advised-edit {
-    margin: 0;
-    height: 1.5rem;
-    flex-grow: 0;
-}
-
-.club-popup-name-input {
-    font-size: 1.1rem;
-    margin-right: 1rem;
-    flex-grow: 1;
-}
-
-.club-popup-description-input {
-    width: 97%;
-    height: 6rem;
-    margin: 0.25rem 0 0.5rem;
-    padding: 0.25rem;
-
-    font-family: $alt-font;
-    font-size: 0.45rem;
-}
-
-.club-popup-link-input,
-.club-popup-link-label {
-    font-size: 0.75rem;
-}
-
-.club-popup-link-input {
-    width: 60%;
-}
-
-/*
- * ###################################
- * ##### EXEC AND COMMITTEE EDIT #####
- * ###################################
- */
-
-.club-popup-card-edit-heading {
-    margin: 1rem 0 0.25rem;
-    font-size: 1rem;
-    text-align: center;
-    color: $text-light;
-}
-
-// Override bottom margins
-.club-popup .exec-edit,
-.club-popup .committee-edit {
-    margin-bottom: 0.5rem;
-}
-
-.club-popup-card-edit-list {
-    display: flex;
-    flex-direction: column;
-}
-
-.club-popup-add-container {
-    text-align: center;
-}
-
-/*
- * #####################################
- * ##### SUBMIT AND DELETE BUTTONS #####
- * #####################################
- */
-
-.club-popup-action-button-box {
-    margin-top: 0.5rem;
-    align-items: center;
-}
-
-.club-popup-trash-group {
-    height: fit-content;
-    flex-grow: 1;
-    display: flex;
-}
-
-.club-popup-trash-group.inactive {
-    display: none !important;
-}
-
-.club-popup-delete-text {
-    margin-top: 0.25rem;
-    margin-left: 0.25rem;
-    transition: 0.2s;
-}
-
-.club-popup-delete-box {
-    display: flex;
-    cursor: pointer;
-}
-
-.club-popup-delete-box:hover .club-popup-delete-text {
-    color: $bad;
-}
-
-.club-popup-delete-box:hover .trash-can-path {
-    fill: $bad;
-}
-
-.club-popup-action-button {
-    margin-top: 0 !important;
-    flex-grow: 0;
 }
 
 /*
