+++ client/src/components/edit/committee-edit.scss
@@ -7,7 +7,8 @@
 
 .committee-edit {
     width: 96.5%;
-    padding: 0.25rem 0 0 0.55rem;
+    padding: 0.5rem;
+    margin-bottom: 0.5rem;
     box-shadow: 0 0 0.2rem $background-tint;
 }
