+++ client/src/components/edit/edit-clubs.scss
@@ -1,1 +1,147 @@
 @import '../../custom.scss';
+
+.edit-clubs {
+    text-align: left;
+    margin: 1rem 2rem 0;
+}
+
+.edit-clubs-image {
+    width: 36rem;
+    height: 15rem;
+    margin: auto;
+    display: block;
+    object-fit: cover;
+
+    background-color: $background-tint;
+    color: $background-tint;
+}
+
+// Container is relative because absolute needs a non-static positioned ancestor
+.edit-clubs-image-container {
+    position: relative;
+    width: 100%;
+    height: 15rem;
+}
+
+// Input (button) of type 'file' to upload images
+.edit-clubs-cover-photo-upload {
+    position: absolute;
+    bottom: 0.5rem;
+    left: 2rem;
+}
+
+// Everything underneath the cover photo upload
+// This container has a padding that applies to all the 'edit' elements
+.edit-clubs-edit-bottom {
+    padding: 0.75rem;
+}
+
+.edit-clubs-name-advised-div {
+    display: flex;
+    flex-direction: row;
+    justify-content: left;
+}
+
+.edit-clubs-advised-edit {
+    margin: 0;
+    height: 1.5rem;
+    flex-grow: 0;
+}
+
+.edit-clubs-name-input {
+    font-size: 1.1rem;
+    margin-right: 1rem;
+    flex-grow: 1;
+}
+
+.edit-clubs-description-input {
+    width: 97%;
+    height: 6rem;
+    margin: 0.25rem 0 0.5rem;
+    padding: 0.25rem;
+
+    font-family: $alt-font;
+    font-size: 0.65rem;
+}
+
+.edit-clubs-link-input,
+.edit-clubs-link-label {
+    font-size: 0.75rem;
+}
+
+.edit-clubs-link-input {
+    width: 60%;
+}
+
+/*
+ * ###################################
+ * ##### EXEC AND COMMITTEE EDIT #####
+ * ###################################
+ */
+
+.edit-clubs-card-edit-heading {
+    margin: 1rem 0 0.25rem;
+    font-size: 1rem;
+    text-align: center;
+    color: $text-light;
+}
+
+// Override bottom margins
+.club-popup .exec-edit,
+.club-popup .committee-edit {
+    margin-bottom: 0.5rem;
+}
+
+.edit-clubs-card-edit-list {
+    display: flex;
+    flex-direction: column;
+}
+
+.edit-clubs-add-container {
+    text-align: center;
+}
+
+/*
+ * #####################################
+ * ##### SUBMIT AND DELETE BUTTONS #####
+ * #####################################
+ */
+
+.edit-clubs-action-button-box {
+    margin-top: 0.5rem;
+    align-items: center;
+}
+
+.edit-clubs-trash-group {
+    height: fit-content;
+    flex-grow: 1;
+    display: flex;
+}
+
+.edit-clubs-trash-group.inactive {
+    display: none !important;
+}
+
+.edit-clubs-delete-text {
+    margin-top: 0.25rem;
+    margin-left: 0.25rem;
+    transition: 0.2s;
+}
+
+.edit-clubs-delete-box {
+    display: flex;
+    cursor: pointer;
+}
+
+.edit-clubs-delete-box:hover .edit-clubs-delete-text {
+    color: $bad;
+}
+
+.edit-clubs-delete-box:hover .trash-can-path {
+    fill: $bad;
+}
+
+.edit-clubs-action-button {
+    margin-top: 0 !important;
+    flex-grow: 0;
+}
