+++ client/src/components/edit/edit-events.js
@@ -96,7 +96,7 @@ class EditEvents extends React.Component {
         // Get response and send to user
         if (res.status === 200) {
             alert(`Successfully ${this.state.new ? 'added' : 'edited'} event!`);
-            this.props.parentHistory.push('/');
+            this.props.parentHistory.push(`/events${window.location.search}`);
         } else alert(`${this.state.new ? 'Adding' : 'Editing'} event failed :(`);
     };
