+++ client/src/components/edit/exec-edit.scss
@@ -10,7 +10,8 @@
     width: 99%;
     flex-direction: row;
     display: flex;
-    padding-top: 0.1rem;
+    padding: 0.5rem;
+    margin-bottom: 0.5rem;
     box-shadow: 0 0 0.2rem $background-tint;
 }
 
@@ -76,7 +77,6 @@
 }
 
 .exec-edit-position-input {
-    margin: 0 !important;
     width: 70%;
     font-size: 0.65rem !important;
 }
