+++ client/src/components/edit/image-upload.scss
